+++ src/Spago/PscPackage.hs
@@ -62,7 +62,7 @@ dhallToJSON inputPath outputPath = do
 
   dhall <- readTextFile $ T.fromText inputPath
 
-  jsonVal <- liftIO $ Dhall.JSON.codeToValue Dhall.JSON.NoConversion Dhall.JSON.ForbidWithinJSON inputPath dhall
+  jsonVal <- liftIO $ Dhall.JSON.codeToValue Dhall.JSON.NoConversion Dhall.JSON.ForbidWithinJSON (Just $ Text.unpack inputPath) dhall
 
   writeTextFile outputPath
     $ Text.decodeUtf8
@@ -73,15 +73,15 @@ dhallToJSON inputPath outputPath = do
 -- | Generates a local `packages.json` from the current `packages.dhall`
 insDhall :: Spago m => m ()
 insDhall = do
-  isProject <- testfile PackageSet.path
+  isProject <- testfile PackageSet.packagesPath
   unless isProject $
     die "Missing packages.dhall file. Run `spago psc-package-local-setup` first."
   mktree pscPackageBasePath
   T.touch packagesJsonPath
 
-  PackageSet.ensureFrozen
+  PackageSet.ensureFrozen PackageSet.packagesPath
 
-  try (dhallToJSON PackageSet.path packagesJsonText) >>= \case
+  try (dhallToJSON PackageSet.packagesPath packagesJsonText) >>= \case
     Right _ -> do
       echo $ "Wrote packages.json to " <> packagesJsonText
       echo "Now you can run `psc-package install`."