+++ CHANGELOG.md
@@ -6,12 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
 Breaking changes (!!!):
 - **Flags and arguments that you want to give to `purs` are now passed with `--purs-args`**
 
   The previous behaviour in which all arguments that could not parse as `spago` arguments
   were passed along to `purs` was sometimes confusing (e.g. when using `--path` and multiple
   arguments).
+- New `--no-search` flag for `spago docs` to skip patching the documentation using `purescript-docs-search`.
 
 Bugfixes:
 - Do not compile files twice when using `--watch` and Vim (#346)
@@ -21,6 +23,14 @@ Bugfixes:
 - Fix failure to copy to global cache on a different filesystem (#385)
 - Fix watch function on Windows (issue with paths) (#387, #380)
 - "Quit" command in watch mode now actually quits (#390, #389)
+- Look up remote imports dynamically when doing frozen check (#349)
+
+New features:
+- Support watching js files (#407, #205)
+
+Other Improvements:
+- Make no-op `spago install` faster (#409)
+
 
 ## [0.9.0] - 2019-07-30